+++ test/src/image_name_test.dart
@@ -7,19 +7,19 @@ void main() {
     expect(imageName.registry, isNull);
     expect(imageName.repository, isNull);
     expect(imageName.name, equals('ubuntu'));
-    expect(imageName.tag, equals('latest'));
+    expect(imageName.tag, isNull);
 
     imageName = ImageName.fromName('canonical/ubuntu');
     expect(imageName.registry, isNull);
     expect(imageName.repository, equals('canonical'));
     expect(imageName.name, equals('ubuntu'));
-    expect(imageName.tag, equals('latest'));
+    expect(imageName.tag, isNull);
 
     imageName = ImageName.fromName('docker.io/canonical/ubuntu');
     expect(imageName.name, equals('ubuntu'));
     expect(imageName.repository, equals('canonical'));
     expect(imageName.registry, equals('docker.io'));
-    expect(imageName.tag, equals('latest'));
+    expect(imageName.tag, isNull);
 
     imageName = ImageName.fromName('docker.io/canonical/ubuntu:latest');
     expect(imageName.name, equals('ubuntu'));
@@ -43,6 +43,6 @@ void main() {
     expect(imageName.name, equals('ubuntu'));
     expect(imageName.repository, isNull);
     expect(imageName.registry, equals('docker.io'));
-    expect(imageName.tag, equals('latest'));
+    expect(imageName.tag, isNull);
   });
 }